Fix warnings
Summary:
We need to stay on top of these.

Also, we should either disable "declaration shadows a field" warnings or stop shadowing fields!!

Test Plan: n/a

Reviewers: mike.naquin

Reviewed By: mike.naquin

Subscribers: #engineering

Differential Revision: https://phabricator.watchonair.tv/D1241

[This commit was only partially migrated from livelib commit 2cfff97d248af37cf9666001b5a6373e2c50bd35.]

+++ okui/include/onair/okui/Shader.h
@@ -64,12 +64,12 @@ public:
         _transformation.transform(p[1].x, p[1].y, &pT[1].x, &pT[1].y);
         _transformation.transform(p[2].x, p[2].y, &pT[2].x, &pT[2].y);
 
-        _triangle.a.x  = pT[0].x;
-        _triangle.a.y  = pT[0].y;
-        _triangle.b.x  = pT[1].x;
-        _triangle.b.y  = pT[1].y;
-        _triangle.c.x  = pT[2].x;
-        _triangle.c.y  = pT[2].y;
+        _triangle.a.x  = static_cast<GLfloat>(pT[0].x);
+        _triangle.a.y  = static_cast<GLfloat>(pT[0].y);
+        _triangle.b.x  = static_cast<GLfloat>(pT[1].x);
+        _triangle.b.y  = static_cast<GLfloat>(pT[1].y);
+        _triangle.c.x  = static_cast<GLfloat>(pT[2].x);
+        _triangle.c.y  = static_cast<GLfloat>(pT[2].y);
 
         _processTriangle(p, pT, curve);
